+++ src/storage-blob-preview/azext_storage_blob_preview/tests/latest/test_storage_blob_live_scenarios.py
@@ -8,6 +8,7 @@ from datetime import datetime, timedelta
 from azure.cli.testsdk import (LiveScenarioTest, ResourceGroupPreparer, StorageAccountPreparer,
                                JMESPathCheck, JMESPathCheckExists, NoneCheck, api_version_constraint)
 from azure.cli.core.profiles import ResourceType
+from ..storage_test_util import StorageScenarioMixin
 
 
 @api_version_constraint(ResourceType.MGMT_STORAGE, min_api='2016-12-01')
@@ -85,3 +86,68 @@ class StorageBlobUploadLiveTests(LiveScenarioTest):
             self.assertTrue(os.path.isfile(downloaded), 'The file is not downloaded.')
             self.assertEqual(file_size_kb * 1024, os.stat(downloaded).st_size,
                              'The download file size is not right.')
+
+
+class StorageBlobURLScenarioTest(StorageScenarioMixin, LiveScenarioTest):
+    @ResourceGroupPreparer(name_prefix='clitest')
+    @StorageAccountPreparer(kind='StorageV2', name_prefix='clitest', location='eastus2euap')
+    def test_storage_blob_url_scenarios(self, resource_group, storage_account):
+        account_info = self.get_account_info(resource_group, storage_account)
+        container = self.create_container(account_info, prefix="con1")
+
+        local_file1 = self.create_temp_file(128)
+        local_file2 = self.create_temp_file(64)
+        blob_name1 = "/".join(["dir", self.create_random_name(prefix='blob', length=24)])
+
+        # set delete-policy to enable soft-delete
+        self.storage_cmd('storage blob service-properties delete-policy update --enable true --days-retained 2',
+                         account_info)
+        self.storage_cmd('storage blob service-properties delete-policy show',
+                         account_info).assert_with_checks(JMESPathCheck('enabled', True),
+                                                          JMESPathCheck('days', 2))
+        # Prepare blob
+        self.storage_cmd('storage blob upload -c {} -f "{}" -n {} ', account_info,
+                         container, local_file1, blob_name1)
+
+        expiry = (datetime.utcnow() + timedelta(hours=1)).strftime('%Y-%m-%dT%H:%MZ')
+        blob_uri = self.storage_cmd('storage blob generate-sas -n {} -c {} --expiry {} --permissions '
+                                    'rwad --https-only --full-uri -o tsv',
+                                    account_info, blob_name1, container, expiry).output.strip()
+
+        self.cmd('storage blob exists --blob-url {}'.format(blob_uri), checks=JMESPathCheck('exists', True))
+        self.cmd('storage blob show --blob-url {}'.format(blob_uri), checks=[
+            JMESPathCheck('name', blob_name1),
+            JMESPathCheck('properties.contentLength', 128 * 1024),
+            JMESPathCheck('properties.blobTier', 'Hot')])
+
+        self.cmd('storage blob upload -f "{}" --blob-url {} --overwrite'.format(local_file2, blob_uri))
+
+        self.cmd('storage blob show --blob-url {}'.format(blob_uri), checks=[
+            JMESPathCheck('name', blob_name1),
+            JMESPathCheck('properties.contentLength', 64 * 1024)])
+        local_dir = self.create_temp_dir()
+        downloaded = os.path.join(local_dir, 'test.file')
+        self.cmd('storage blob download --blob-url {} -f "{}"'.format(blob_uri, downloaded))
+        self.assertTrue(os.path.isfile(downloaded), 'The file is not downloaded.')
+        self.assertEqual(64 * 1024, os.stat(downloaded).st_size,
+                         'The download file size is not right.')
+
+        self.cmd('storage blob set-tier --blob-url {} --tier Cool'.format(blob_uri))
+        self.cmd('storage blob show --blob-url {}'.format(blob_uri), checks=[
+            JMESPathCheck('name', blob_name1),
+            JMESPathCheck('properties.contentLength', 64 * 1024),
+            JMESPathCheck('properties.blobTier', "Cool")])
+
+        self.cmd('storage blob snapshot --blob-url {}'.format(blob_uri),
+                 checks=JMESPathCheckExists('snapshot'))
+        self.storage_cmd('storage blob list -c {}', account_info, container)\
+            .assert_with_checks(JMESPathCheck('length(@)', 1))
+
+        self.cmd('storage blob delete --blob-url {} --delete-snapshots include '.format(blob_uri))
+
+        self.storage_cmd('storage blob list -c {}', account_info, container).assert_with_checks(
+            JMESPathCheck('length(@)', 0))
+
+        self.cmd('storage blob undelete --blob-url {} '.format(blob_uri))
+        self.storage_cmd('storage blob list -c {}', account_info, container).assert_with_checks(
+            JMESPathCheck('length(@)', 1))
